Move UploadProvider to common components 🚚
